Rename counter classes to avoid confusing.
