+++ osu.Game/osu.Game.csproj
@@ -53,10 +53,10 @@
     <Compile Include="Graphics\Containers\OsuLargeComponent.cs" />
     <Compile Include="Graphics\Processing\RatioAdjust.cs" />
     <Compile Include="Graphics\TextAwesome.cs" />
-    <Compile Include="Graphics\UserInterface\Counter.cs" />
-    <Compile Include="Graphics\UserInterface\KeyBoardCount.cs" />
     <Compile Include="Graphics\UserInterface\KeyCounter.cs" />
-    <Compile Include="Graphics\UserInterface\MouseCount.cs" />
+    <Compile Include="Graphics\UserInterface\KeyCounterKeyBoard.cs" />
+    <Compile Include="Graphics\UserInterface\KeyCounterCollection.cs" />
+    <Compile Include="Graphics\UserInterface\KeyCounterMouse.cs" />
     <Compile Include="Online\API\APIAccess.cs" />
     <Compile Include="Online\API\APIRequest.cs" />
     <Compile Include="Online\API\OAuth.cs" />
